+++ spec/dummy_app/app/models/user.rb
@@ -5,6 +5,7 @@ class User < ApplicationRecord
     :recoverable, :rememberable, :validatable,
     :castle_protectable, castle_hooks: {
       after_login: true,
+      after_password_reset_request: true,
       before_registration: true
     }
 end