+++ spec/spec_helper.rb
@@ -41,6 +41,7 @@ RSpec.configure do |config|
 
   config.expect_with :rspec do |c|
     c.syntax = :expect
+    c.max_formatted_output_length = nil
   end
 
   config.after do